+++ src/Sentry.Unity/SentryUnityOptions.cs
@@ -35,7 +35,7 @@ namespace Sentry.Unity
         /// <summary>
         /// Whether the SDK should be in <see cref="Debug"/> mode only while in the Unity Editor.
         /// </summary>
-        public bool DebugOnlyInEditor { get; set; } = true;
+        public bool DebugOnlyInEditor { get; set; }
 
         private CompressionLevelWithAuto _requestBodyCompressionLevel = CompressionLevelWithAuto.Auto;