+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- New config window ([#243](https://github.com/getsentry/sentry-unity/pull/243))
+
 ## 0.3.3
 
 ### Fixes